# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Online Competitive Exam Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RegisterRequest {
  fullName: string;
  username: string;
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  STUDENT: 'STUDENT',
  ADMIN: 'ADMIN',
} as const;

export interface User {
  id: number;
  fullName: string;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Exam {
  id: number;
  title: string;
  description?: string;
  examType: string;
  duration: number;
  totalMarks: number;
  negativeMarks: number;
  isPublished: boolean;
  questionCount?: number;
  attemptCount?: number;
  createdAt: string;
}

export interface Section {
  id: number;
  examId: number;
  sectionName: string;
  sectionDuration: number;
  questionCount?: number;
}

export type QuestionQuestionType = typeof QuestionQuestionType[keyof typeof QuestionQuestionType];


export const QuestionQuestionType = {
  MCQ: 'MCQ',
  TRUE_FALSE: 'TRUE_FALSE',
  NUMERICAL: 'NUMERICAL',
  ASSERTION_REASON: 'ASSERTION_REASON',
} as const;

export interface Question {
  id: number;
  examId: number;
  sectionId?: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  explanation?: string;
  marks: number;
  negativeMarks: number;
  questionType: QuestionQuestionType;
}

export type SectionWithQuestions = Section & {
  questions?: Question[];
};

export type ExamDetail = Exam & {
  sections?: SectionWithQuestions[];
};

export interface ExamListResponse {
  exams: Exam[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateExamRequest {
  title: string;
  description?: string;
  examType: string;
  duration: number;
  totalMarks: number;
  negativeMarks: number;
}

export interface UpdateExamRequest {
  title?: string;
  description?: string;
  examType?: string;
  duration?: number;
  totalMarks?: number;
  negativeMarks?: number;
}

export interface CreateSectionRequest {
  sectionName: string;
  sectionDuration: number;
}

export interface QuestionListResponse {
  questions: Question[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateQuestionRequestQuestionType = typeof CreateQuestionRequestQuestionType[keyof typeof CreateQuestionRequestQuestionType];


export const CreateQuestionRequestQuestionType = {
  MCQ: 'MCQ',
  TRUE_FALSE: 'TRUE_FALSE',
  NUMERICAL: 'NUMERICAL',
  ASSERTION_REASON: 'ASSERTION_REASON',
} as const;

export interface CreateQuestionRequest {
  examId: number;
  sectionId?: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  explanation?: string;
  marks: number;
  negativeMarks: number;
  questionType: CreateQuestionRequestQuestionType;
}

export interface UpdateQuestionRequest {
  questionText?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: string;
  explanation?: string;
  marks?: number;
  negativeMarks?: number;
}

export interface StartAttemptRequest {
  examId: number;
}

export interface SaveAnswerRequest {
  questionId: number;
  selectedOption?: string;
  markedForReview?: boolean;
}

export interface AnswerSavedResult {
  success: boolean;
}

export interface SavedAnswer {
  questionId: number;
  selectedOption?: string;
  markedForReview?: boolean;
}

export interface AttemptDetail {
  id: number;
  examId: number;
  exam?: ExamDetail;
  startedAt: string;
  timeRemaining: number;
  savedAnswers?: SavedAnswer[];
  isSubmitted: boolean;
}

export interface AttemptSummary {
  id: number;
  examId: number;
  examTitle: string;
  examType: string;
  score?: number;
  accuracy?: number;
  totalCorrect?: number;
  totalWrong?: number;
  timeTaken?: number;
  submittedAt: string;
  rank?: number;
}

export interface SectionResult {
  sectionId: number;
  sectionName: string;
  correct: number;
  wrong: number;
  unattempted: number;
  score: number;
  accuracy: number;
}

export interface QuestionResult {
  questionId: number;
  questionText: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer: string;
  selectedOption?: string;
  isCorrect: boolean;
  explanation?: string;
  sectionName?: string;
}

export interface AttemptResult {
  id: number;
  examId: number;
  examTitle: string;
  score: number;
  totalMarks: number;
  accuracy: number;
  totalCorrect: number;
  totalWrong: number;
  totalUnattempted: number;
  timeTaken: number;
  rank?: number;
  negativeMarks?: number;
  sectionResults?: SectionResult[];
  questionResults?: QuestionResult[];
}

export interface SubjectPerformance {
  subject: string;
  accuracy: number;
  totalQuestions: number;
  correctQuestions: number;
}

export interface ScoreHistory {
  date: string;
  score: number;
  examTitle: string;
}

export interface StudentAnalytics {
  totalAttempts: number;
  averageScore: number;
  averageAccuracy: number;
  bestScore: number;
  totalTimeSpent: number;
  subjectPerformance?: SubjectPerformance[];
  recentAttempts?: AttemptSummary[];
  weakAreas?: string[];
  strongAreas?: string[];
  scoreHistory?: ScoreHistory[];
}

export interface AdminStats {
  totalUsers: number;
  totalExams: number;
  totalQuestions: number;
  activeExams: number;
  totalAttempts: number;
  dailyActiveUsers: number;
  publishedExams?: number;
  draftExams?: number;
}

export interface ExamAnalytic {
  examId: number;
  examTitle: string;
  attemptCount: number;
  avgScore: number;
}

export interface QuestionAnalytic {
  questionId: number;
  questionText: string;
  wrongRate: number;
}

export interface DayActivity {
  date: string;
  activeUsers: number;
  attempts: number;
}

export interface ExamTypeDistribution {
  examType: string;
  count: number;
}

export interface AdminAnalytics {
  mostAttemptedExams: ExamAnalytic[];
  hardestQuestions: QuestionAnalytic[];
  userActivityByDay: DayActivity[];
  completionRate: number;
  examTypeDistribution?: ExamTypeDistribution[];
}

export interface PdfUploadRequest {
  file?: Blob;
  examId?: number;
  sectionId?: number;
}

export interface PdfExtractResponse {
  extractedCount: number;
  questions: Question[];
  examId?: number;
}

export interface ToggleBookmarkRequest {
  questionId: number;
}

export interface BookmarkResult {
  bookmarked: boolean;
}

export type GetExamsParams = {
examType?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetQuestionsParams = {
examId?: number;
sectionId?: number;
page?: number;
limit?: number;
};

export type GetUsersParams = {
page?: number;
limit?: number;
search?: string;
};