# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Online Competitive Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/upload-pdf:
    post:
      operationId: uploadPdf
      tags: [admin]
      summary: Upload a PDF and auto-extract questions (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PdfUploadRequest"
      responses:
        "200":
          description: Extracted questions preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfExtractResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get platform-wide analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
  /admin/users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{userId}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [STUDENT, ADMIN]
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, username, role, createdAt]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
...
    Question:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        sectionId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        marks:
          type: integer
        negativeMarks:
          type: number
        questionType:
          type: string
          enum: [MCQ, TRUE_FALSE, NUMERICAL, ASSERTION_REASON]
      required: [id, examId, questionText, optionA, optionB, optionC, optionD, correctAnswer, marks, negativeMarks, questionType]
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalExams:
          type: integer
        totalQuestions:
          type: integer
        activeExams:
          type: integer
        totalAttempts:
          type: integer
        dailyActiveUsers:
          type: integer
        publishedExams:
          type: integer
        draftExams:
          type: integer
      required: [totalUsers, totalExams, totalQuestions, activeExams, totalAttempts, dailyActiveUsers]
    AdminAnalytics:
      type: object
      properties:
        mostAttemptedExams:
          type: array
          items:
            $ref: "#/components/schemas/ExamAnalytic"
        hardestQuestions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionAnalytic"
        userActivityByDay:
          type: array
          items:
            $ref: "#/components/schemas/DayActivity"
        completionRate:
          type: number
        examTypeDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ExamTypeDistribution"
      required: [mostAttemptedExams, hardestQuestions, userActivityByDay, completionRate]
    ExamAnalytic:
      type: object
      properties:
        examId:
          type: integer
        examTitle:
          type: string
        attemptCount:
          type: integer
        avgScore:
          type: number
      required: [examId, examTitle, attemptCount, avgScore]
    QuestionAnalytic:
      type: object
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        wrongRate:
          type: number
      required: [questionId, questionText, wrongRate]
    DayActivity:
      type: object
      properties:
        date:
          type: string
        activeUsers:
          type: integer
        attempts:
          type: integer
      required: [date, activeUsers, attempts]
    ExamTypeDistribution:
      type: object
      properties:
        examType:
          type: string
        count:
          type: integer
      required: [examType, count]
    PdfUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
        examId:
          type: integer
        sectionId:
          type: integer
    PdfExtractResponse:
      type: object
      properties:
        extractedCount:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        examId:
          type: integer
      required: [extractedCount, questions]
...