# lib/api-zod/src/generated/types/adminAnalytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Online Competitive Exam Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DayActivity } from './dayActivity';
import type { ExamAnalytic } from './examAnalytic';
import type { ExamTypeDistribution } from './examTypeDistribution';
import type { QuestionAnalytic } from './questionAnalytic';

export interface AdminAnalytics {
  mostAttemptedExams: ExamAnalytic[];
  hardestQuestions: QuestionAnalytic[];
  userActivityByDay: DayActivity[];
  completionRate: number;
  examTypeDistribution?: ExamTypeDistribution[];
}