# lib/api-zod/src/generated/types/examTypeDistribution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Online Competitive Exam Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExamTypeDistribution {
  examType: string;
  count: number;
}